callibri-export/requirements.txt
requests
python-dotenv
openpyxl
customtkinter
gspread
google-auth
